+++ src/main/java/com/scc/service/DogService.java
@@ -21,6 +21,7 @@ import com.scc.model.PgArbreGenealogie;
 import com.scc.model.PgDog;
 import com.scc.repository.PgArbreRepository;
 import com.scc.repository.PgDogRepository;
+import com.scc.repository.DogRepository;
 
 import io.quarkus.security.identity.SecurityIdentity;
 
@@ -34,13 +35,20 @@ public class DogService extends AbstractGenericService<Dog> {
     PgArbreRepository pgArbreRepository;
 
     @Inject 
+    DogRepository dogRepository;
+
+    @Inject 
     PgDogRepository pgDogRepository;
 
     @Inject
     SecurityIdentity identity;
     
-    public void save(PgDog _dog) {
-        //PgDogRepository.persistAndFlush(_dog);
+    public void populateDog(PgDog _dog) {
+        pgDogRepository.persistAndFlush(_dog);
+    }
+
+    public void populateParents(PgArbreGenealogie _parents) {
+        pgArbreRepository.persistAndFlush(_parents);
     }
 
     public Dog findDogById(Integer id) throws IllegalArgumentException, IllegalAccessException, ClassNotFoundException{
@@ -54,7 +62,7 @@ public class DogService extends AbstractGenericService<Dog> {
         Dog _dog = new Dog();
         try {
 
-             _dogs = pgDogRepository.findByToken(token);
+             _dogs = dogRepository.findByToken(token);
              PgArbreGenealogie _parents = null;
              
              if (_dogs.size()>0 && _dogs != null) {
@@ -91,11 +99,11 @@ public class DogService extends AbstractGenericService<Dog> {
         
         if (_idFather != null) {
             System.out.println("findFamilyTree level " + _level + "Father ["+_idFather+"]");
-            _f = handleRole(pgDogRepository.findDog(_idFather));
+            _f = handleRole(dogRepository.findDog(_idFather));
         }
         if (_idMother != null) {
             System.out.println("findFamilyTree level " + _level + "Mother ["+_idMother+"]");
-            _m = handleRole(pgDogRepository.findDog(_idMother));
+            _m = handleRole(dogRepository.findDog(_idMother));
         }
         if (_m != null) {
             _dog.setMother(_m);
