+++ src/main/resources/application.properties
@@ -24,8 +24,8 @@ quarkus.native.enable-https-url-handler=true
 %prod.quarkus.keycloak.policy-enforcer.enable=true
 
 # Disables policy enforcement for a path
-#quarkus.keycloak.policy-enforcer.paths.1.path=/documentation
-#quarkus.keycloak.policy-enforcer.paths.1.enforcement-mode=DISABLED
+quarkus.keycloak.policy-enforcer.paths.1.path=/public/*
+quarkus.keycloak.policy-enforcer.paths.1.enforcement-mode=DISABLED
 
 # Database
 quarkus.hibernate-orm.database.generation=drop-and-create
@@ -62,9 +62,9 @@ quarkus.index-dependency.hibernate.group-id=org.hibernate
 quarkus.index-dependency.hibernate.artifact-id=hibernate-core
 
 # swagger
-quarkus.swagger-ui.path=/documentation
+quarkus.swagger-ui.path=/public/documentation
 # default endpoint location /openapi
-#quarkus.smallrye-openapi.path=/swagger
+quarkus.smallrye-openapi.path=/public/openapi
 # make it available in production
 quarkus.swagger-ui.always-include=true
